pyproject.toml
[project]
name = "agentic-rag-with-langgraph"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "jupyter>=1.1.1",
    "langchain>=0.3.26",
    "langchain-chroma>=0.2.4",
    "langchain-community>=0.3.27",
    "langchain-openai>=0.3.27",
    "langgraph>=0.5.2",
    "pandas>=2.3.1",
    "pydantic>=2.11.7",
    "python-dotenv>=1.1.1",
    "tavily-python>=0.7.9",
    "tqdm>=4.67.1",
]
